refactor(cli): extract BookPlacementCommands from BookCommands
- Move `addToShelf` logic to the newly created `BookPlacementCommands` class.
- Remove redundant `addToShelf` method and related imports from `BookCommands`.
- Improve separation of concerns by isolating book placement-related commands into a dedicated class.
